requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
frozenlist==1.3.0
httpagentparser==1.9.3
idna==3.3
maxminddb==2.2.0
multidict==6.0.2
pycodestyle==2.8.0
python-dotenv==0.20.0
requests==2.28.1
toml==0.10.2
urllib3==1.26.10
yarl==1.7.2
